# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: ANZH Beauty API
  version: '1.0.0'
  description: |
    Бэкенд мини-приложения ANZH Cosmetology.

    **Личность.** Все запросы, кроме витрины и здоровья, требуют
    сессии. Сессия выдаётся в обмен на `initData` из Telegram, подпись
    которого проверяется на сервере (HMAC + срок жизни 1 час). Роль
    берётся с сервера по `telegram_id`, а не из запроса.

    **Деньги.** Целое в минорных единицах плюс код валюты: `45000` и
    `GEL` означают 450.00 лари. Telegram Stars — отдельное поле
    `stars_amount`, это не валюта ISO-4217.

    **Время.** Всё в UTC, ISO-8601. Часовой пояс салона —
    `Asia/Tbilisi` — параметр конфигурации, а не свойство данных.

    **Итог визита считает сервер.** Клиент присылает состав, а не сумму.
servers:
  - url: '{base}/api/v1'
    description: Базовый адрес берётся из переменной окружения клиента
    variables:
      base:
        default: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      tags: [Каталог]
      summary: Витрина
      description: Открыта без сессии — каталог публичен. Снятые с витрины процедуры не показываются.
      security: []
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [procedure, training, analysis] }
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items: { $ref: '#/components/schemas/Service' }
        '422':
          description: '`bad_kind` — неизвестный вид. Пустой список на опечатку ищут в данных, а не в коде'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /services/{slug}:
    parameters:
      - $ref: '#/components/parameters/Slug'
    get:
      tags: [Каталог]
      summary: Одна процедура
      security: []
      responses:
        '200':
          description: Процедура
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Каталог]
      summary: Правка прайса
      description: |
        Только владелица. Отдельной админки нет — прайс правится из
        кабинета. Обновляются только присланные поля: `null` стирает
        значение, отсутствие поля оставляет как было.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServicePatch' }
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Invalid' }
    delete:
      tags: [Каталог]
      summary: Снять с витрины
      description: |
        Мягко: строка остаётся, на неё ссылается история визитов.
        Жёсткое удаление сломало бы карточки прошлых клиенток.
      responses:
        '200':
          description: Снято
          content:
            application/json:
              schema:
                type: object
                properties: { archived: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /services/{slug}/restore:
    parameters:
      - $ref: '#/components/parameters/Slug'
    post:
      tags: [Каталог]
      summary: Вернуть на витрину
      responses:
        '200':
          description: Возвращено
          content:
            application/json:
              schema:
                type: object
                properties: { restored: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /services/quote:
    post:
      tags: [Каталог]
      summary: Сколько выйдет визит
      description: |
        Считает ту же функция, что и при создании заявки — поэтому
        показанный итог и записанный совпадают по построению.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slugs]
              properties:
                slugs:
                  type: array
                  items: { type: string }
                  minItems: 1
                  maxItems: 6
      responses:
        '200':
          description: Расчёт
          content:
            application/json:
              schema:
                type: object
                properties:
                  quote: { $ref: '#/components/schemas/Quote' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: |
            `empty_visit` · `too_many` · `unknown_service` ·
            `incomplete_service` (нет цены или длительности) ·
            `mixed_currency`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /clients:
    get:
      tags: [Клиентки]
      summary: Список клиенток с историей
      description: Мастер видит тех, кто приходил к нему; владелица — всех.
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items: { $ref: '#/components/schemas/ClientSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /clients/{telegramId}:
    get:
      tags: [Клиентки]
      summary: Карточка клиентки
      description: |
        История визитов, неявки, потраченное, заработанная скидка.

        🔒 **Анкеты здоровья и результатов разборов здесь нет.** Они
        приезжают отдельным эндпоинтом, с записью в журнал доступа и
        после ответов юриста по §1 контракта. Иначе каждое открытие
        списка клиенток становилось бы доступом к медданным.

        Мастеру, к которому эта клиентка не приходила, отвечаем `404`,
        а не `403`: `403` подтвердил бы, что такая клиентка есть.
      parameters:
        - { name: telegramId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Карточка и история
          content:
            application/json:
              schema:
                type: object
                properties:
                  client: { $ref: '#/components/schemas/ClientCard' }
                  history:
                    type: array
                    items: { $ref: '#/components/schemas/HistoryEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Invalid' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <токен>`. Полезная нагрузка ровно такая:
        `{ telegram_id, app_source: "beauty", role }`. Поле `app_source`
        сверяется при каждом запросе — токен другого приложения это не
        открывает.
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]([a-z0-9-]*[a-z0-9])?$' }
...
  responses:
    Unauthorized:
      description: Нет сессии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: no_session, message: Нужно войти заново } }
...
    Forbidden:
      description: Роли или прав на объект не хватает
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Объекта нет
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Invalid:
      description: Не проходит проверку
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Единая форма на все отказы. Ни трассировки стека, ни текста
        запроса к базе: по ним читается устройство сервера, а в тексте
        запроса могут оказаться данные из анкеты.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: slot_taken }
            message: { type: string, example: Время уже занято }
...
    Service:
      type: object
      properties:
        slug: { type: string }
        kind: { type: string, enum: [procedure, training, analysis] }
        category:
          type: string
          enum: [inj, clean, peel, apparatus, care, derma, brows, training, analysis]
        zone: { type: string, enum: [face, body, brows, general] }
        title:
          type: object
          properties: { ru: { type: string }, en: { type: string } }
        subtitle:
          type: object
          properties: { ru: { type: [string, 'null'] }, en: { type: [string, 'null'] } }
        description:
          type: object
          properties: { ru: { type: [string, 'null'] }, en: { type: [string, 'null'] } }
        price_minor:
          type: [integer, 'null']
          format: int64
          description: Минорные единицы. 45000 при currency GEL — это 450.00 лари.
        currency: { type: [string, 'null'], pattern: '^[A-Z]{3}$' }
        price_from: { type: boolean, description: Цена «от» — объём заранее не известен }
        stars_amount:
          type: [integer, 'null']
          description: Telegram Stars. Не валюта ISO-4217, с выручкой салона не смешивается.
        duration_min: { type: [integer, 'null'] }
        days: { type: [integer, 'null'], description: Для обучения }
        includes: { type: array, items: { type: string } }
        contraindications:
          type: array
          items: { type: string }
          description: Сверяются с анкетой клиентки перед подтверждением записи.
        photo_url: { type: [string, 'null'] }
        sort: { type: integer }
    ServicePatch:
      type: object
      minProperties: 1
      description: Только присланные поля. `null` стирает значение.
      properties:
        title_ru: { type: string }
        title_en: { type: string }
        subtitle_ru: { type: [string, 'null'] }
        subtitle_en: { type: [string, 'null'] }
        description_ru: { type: [string, 'null'] }
        description_en: { type: [string, 'null'] }
        price_minor: { type: [integer, 'null'], format: int64, minimum: 0 }
        currency: { type: [string, 'null'], pattern: '^[A-Z]{3}$' }
        price_from: { type: boolean }
        stars_amount: { type: [integer, 'null'], minimum: 0 }
        duration_min: { type: [integer, 'null'], minimum: 1 }
        days: { type: [integer, 'null'], minimum: 1 }
        includes: { type: array, items: { type: string } }
        contraindications: { type: array, items: { type: string } }
        photo_url: { type: [string, 'null'] }
        sort: { type: integer }
    Quote:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              service_id: { type: string, format: uuid }
              slug: { type: string }
              title_ru: { type: string }
              title_en: { type: string }
              price_minor: { type: integer, format: int64 }
              currency: { type: string }
              duration_min: { type: integer }
        subtotal_minor: { type: integer, format: int64 }
        discount_percent:
          type: integer
          maximum: 30
          description: 10% за задание, не больше 30. Считается из подтверждённых оснований.
        discount_minor: { type: integer, format: int64 }
        total_minor: { type: integer, format: int64 }
        currency: { type: string }
        duration_min: { type: integer }
    BookingStatus:
      type: string
      enum: [pending, confirmed, declined, done, no_show, cancelled]
...
    ClientSummary:
      type: object
      properties:
        telegram_id: { type: integer, format: int64 }
        first_name: { type: [string, 'null'] }
        last_name: { type: [string, 'null'] }
        username: { type: [string, 'null'] }
        visits: { type: integer }
        no_shows: { type: integer }
        cancelled: { type: integer }
        spent_minor: { type: integer, format: int64 }
        currency: { type: [string, 'null'] }
        first_visit: { type: [string, 'null'], format: date-time }
        last_visit: { type: [string, 'null'], format: date-time }
    ClientCard:
      type: object
      properties:
        telegram_id: { type: integer, format: int64 }
        first_name: { type: [string, 'null'] }
        last_name: { type: [string, 'null'] }
        username: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        visits: { type: integer }
        no_shows: { type: integer }
        cancelled: { type: integer }
        spent_minor: { type: integer, format: int64 }
        currency: { type: [string, 'null'] }
        discount_percent: { type: integer }
        health:
          type: object
          description: 🔒 Пока закрыто — см. ОТЧЁТ-§1.md
          properties:
            available: { type: boolean }
            reason: { type: string, example: pending_legal_review }
    HistoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/BookingStatus' }
        total_minor: { type: integer, format: int64 }
        currency: { type: string }
        staff_id: { type: integer, format: int64 }
        services: { type: array, items: { type: string } }